src/E2E-Sample/Voice/ModelConvert/requirements.txt
torch==1.13
pyyaml>=5.1
onnxruntime==1.8.1
typeguard==2.13.3
onnxsim